type fix

+++ tests/join_tests/test_join_4.cpp
@@ -68,8 +68,8 @@ int main(int argc, char *argv[])
     size_t runs = 1;
     size_t stream_len = 0;
     size_t n_keys = 1;
-    size_t lower_bound = 0;
-    size_t upper_bound = 0;
+    int64_t lower_bound = 0;
+    int64_t upper_bound = 0;
     // initalize global variable
     global_sum = 0;
     // arguments from command line
